+++ src/services/ShippingCategories.php
@@ -212,7 +212,7 @@ class ShippingCategories extends Component
         $record = ShippingCategoryRecord::findOne($id);
 
         if ($record) {
-            return (bool)$record->delete();
+            return (bool)$record->softDelete();
         }
 
         // Clear cache