+++ src/services/TaxCategories.php
@@ -272,7 +272,7 @@ class TaxCategories extends Component
         $record = TaxCategoryRecord::findOne($id);
 
         if ($record) {
-            return (bool)$record->delete();
+            return (bool)$record->softDelete();
         }
 
         return false;